% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TMA11.R
\name{TMA11}
\alias{TMA11}
\title{Import Tissue MicroArray files as either HDF5 or Zarr}
\usage{
TMA11(format = c("hdf5", "zarr"))
}
\arguments{
\item{format}{character(1) Either "hdf5" or "zarr" to indicate the desired
file types to obtain from the OSN. HDF5 files will be represented via
the \code{cytomapper::CytoImageList} class and Zarr files will be returned as
a single \code{ZarrRemote} object.}
}
\description{
The function uses \code{ZarrExperiment} to locate remote files. Depending on the
\code{format} input, either a \code{ZarrRemote} class is produced for \code{Zarr} files or a
\code{CytoImageList} is given for \code{HDF5} files. All files are hosted on the Open
Storage Network (OSN) and queried via \code{ExperimentHub} (pending).
}
\examples{

browseURL("https://mghp.osn.xsede.org/bir190004-bucket01/index.html#TMA11/")
browseURL("https://www.synapse.org/#!Synapse:syn22345748/wiki/")

}
\seealso{
\url{https://mghp.osn.xsede.org/bir190004-bucket01/index.html#TMA11/},
\url{https://www.synapse.org/#!Synapse:syn22345748/wiki/}
}
